refactor: move orso providers to generic module
